+++ packages/gatsby-source-wordpress/src/normalize.js
@@ -477,6 +477,7 @@ exports.downloadMediaFiles = async ({
   touchNode,
   getNode,
   _auth,
+  reporter,
 }) =>
   Promise.all(
     entities.map(async e => {
@@ -511,6 +512,7 @@ exports.downloadMediaFiles = async ({
               createNodeId,
               parentNodeId: e.id,
               auth: _auth,
+              reporter,
             })
 
             if (fileNode) {
@@ -644,6 +646,17 @@ exports.createUrlPathsFromLinks = entities =>
       } catch (error) {
         e.path = e.link
       }
+    }
+    return e
+  })
+
+exports.normalizeMenuItems = entities =>
+  entities.map(e => {
+    if (e.__type === `wordpress__menus_menus_items`) {
+      // in case of nested menus items might be object
+      // this converts it into array so it's consistent
+      // and queried in simple manner
+      e.items = _.values(e.items)
     }
     return e
   })