+++ packages/gatsby-transformer-sqip/src/extend-node-type.js
@@ -1,4 +1,5 @@
 const { resolve } = require(`path`)
+const md5File = require(`bluebird`).promisify(require(`md5-file`))
 
 const {
   DuotoneGradientType,
@@ -43,14 +44,16 @@ module.exports = async args => {
 
 async function sqipSharp({ type, cache, getNodeAndSavePathDependency, store }) {
   const program = store.getState().program
-  const cacheDir = resolve(`${program.directory}/.cache/sqip/`)
+  const cacheDir = resolve(
+    `${program.directory}/node_modules/.cache/gatsby-transformer-sqip/`
+  )
 
   await ensureDir(cacheDir)
 
   return {
     sqip: {
       type: new GraphQLObjectType({
-        name: `Sqip`,
+        name: `SqipSharp`,
         fields: {
           svg: { type: GraphQLString },
           dataURI: { type: GraphQLString },
@@ -107,6 +110,7 @@ async function sqipSharp({ type, cache, getNodeAndSavePathDependency, store }) {
         }
 
         const file = getNodeAndSavePathDependency(image.parent, context.path)
+        const { contentDigest } = image.internal
 
         const job = await queueImageResizing({ file, args: sharpArgs })
 
@@ -120,6 +124,7 @@ async function sqipSharp({ type, cache, getNodeAndSavePathDependency, store }) {
         return generateSqip({
           cache,
           cacheDir,
+          contentDigest,
           absolutePath,
           numberOfPrimitives,
           blur,
@@ -138,14 +143,16 @@ async function sqipContentful({ type, cache, store }) {
   const cacheImage = require(`gatsby-source-contentful/cache-image`)
 
   const program = store.getState().program
-  const cacheDir = resolve(`${program.directory}/.cache/sqip/`)
+  const cacheDir = resolve(
+    `${program.directory}/node_modules/.cache/gatsby-transformer-sqip/`
+  )
 
   await ensureDir(cacheDir)
 
   return {
     sqip: {
       type: new GraphQLObjectType({
-        name: `Sqip`,
+        name: `SqipContentful`,
         fields: {
           svg: { type: GraphQLString },
           dataURI: { type: GraphQLString },
@@ -219,9 +226,12 @@ async function sqipContentful({ type, cache, store }) {
 
         const absolutePath = await cacheImage(store, asset, options)
 
+        const contentDigest = await md5File(absolutePath)
+
         return generateSqip({
           cache,
           cacheDir,
+          contentDigest,
           absolutePath,
           numberOfPrimitives,
           blur,