+++ www/src/components/shared/footer-links.js
@@ -18,12 +18,15 @@ const FooterList = styled.ul`
 
   li {
     display: inline-block;
-
-    &:first-of-type:after {
+    &:after {
       color: ${colors.grey[30]};
       content: "•";
       padding-left: 1em;
-      margin-right: 1em;
+      padding-right: 1em;
+    }
+
+    &:last-of-type:after {
+      content: none;
     }
 
     a {
@@ -42,6 +45,9 @@ const FooterLinks = props => (
   <FooterList bottomMargin={props.bottomMargin}>
     <li>
       <Link to="/accessibility-statement">Accessibility Statement</Link>
+    </li>
+    <li>
+      <a href="/guidelines/logo/">Logo &amp; Assets</a>
     </li>
     <li>
       <a href="https://www.gatsbyjs.com">Gatsbyjs.com</a>