+++ packages/gatsby-plugin-guess-js/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-guess-js/compare/gatsby-plugin-guess-js@1.0.7...gatsby-plugin-guess-js@1.1.0) (2019-06-20)
+
+**Note:** Version bump only for package gatsby-plugin-guess-js
+
+## [1.0.7](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-guess-js/compare/gatsby-plugin-guess-js@1.0.6...gatsby-plugin-guess-js@1.0.7) (2019-06-11)
+
+**Note:** Version bump only for package gatsby-plugin-guess-js
+
 ## [1.0.6](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-guess-js/compare/gatsby-plugin-guess-js@1.0.5...gatsby-plugin-guess-js@1.0.6) (2019-04-23)
 
 **Note:** Version bump only for package gatsby-plugin-guess-js