+++ packages/gatsby-telemetry/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-telemetry/compare/gatsby-telemetry@1.1.0...gatsby-telemetry@1.1.1) (2019-06-27)
+
+**Note:** Version bump only for package gatsby-telemetry
+
+# [1.1.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-telemetry/compare/gatsby-telemetry@1.0.12...gatsby-telemetry@1.1.0) (2019-06-20)
+
+**Note:** Version bump only for package gatsby-telemetry
+
+## [1.0.12](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-telemetry/compare/gatsby-telemetry@1.0.11...gatsby-telemetry@1.0.12) (2019-06-18)
+
+**Note:** Version bump only for package gatsby-telemetry
+
 ## [1.0.11](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-telemetry/compare/gatsby-telemetry@1.0.10...gatsby-telemetry@1.0.11) (2019-05-31)
 
 ### Bug Fixes